Mostly done with DES F function.  Code is not operational, but its in the perfect state for testing DES so if i need to recover to healthy state i need to go to the previous snapshot.

+++ ksu/cs7530/obscura/encryption/FeistelFFunction.java
@@ -1,6 +1,8 @@
 package ksu.cs7530.obscura.encryption;
 
+import java.math.BigInteger;
+
 public interface FeistelFFunction {
 
-    public int transform(int input, long key);
+    public BigInteger transform(BigInteger input, long key);
 }
